InteractiveRenderTest : Make less sensitive to sampling strategy
A Renderer may choose not to sample all lights for each camera sample, so renders with a red and a green light may not converge quickly enough to satisfy the assertions in `testAddLight()`. Since we want to test the _addition_ of a light, and we have other tests for lights that exist at the start of the renderer, the simple fix is to just use one light for the test.