OSLObject : Avoid unnecessary resampling of PointsPrimitives
When working with objects that don't fit in the cache, this can have a huge impact on performance. The issue is that pulling on both `inPlug()->objectPlug()` and `resampledInPlug()->objectPlug()` gives us two upstream paths which lead to the source object, and if the cache isn't big enough to hold the source object then we end up with multiple upstream evaluations. Chaining a bunch of OSLObjects leads to exponential explosion of evaluations of the source object.

While the fix here has a worthwhile impact, it doesn't elimate additional upstream computes entirely. The issue is that `OSLObject::hashProcessedObject()` must still call `resampledInPlug()->objectPlug()->hash()` because we don't know if we're dealing with PointsPrimitives at this stage. And `PrimitiveVariableProcessor::hashProcessedObject()` hashes `PrimitiveVariableProcessor::namesPlug()`, which pulls on `OSLObject::resampledNamesPlug(), which in turn pulls on `OSLObject::inPlug()->objectPlug()`. This could be avoided by removing substitutions from `PrimitiveVariableProcessor::namesPlug()`, but we can't do that for the general case.