StandardNodeGadget : Fix problem with ShaderAssignment drag snapping
When dragging from a ShaderAssignment to a Shader with an `out.foo` output, we were making the connection to the `out` plug, when we wanted to make it to `out.foo`. Replacing `<` with `<=` in our comparison gives preference to the nested plug in this situation, and produces the connections we want.