UIEditor : Try to make section naming easier
Trying to grab the focus for the name widget when a new section has been added. Doesn't really work - seems that focus gets grabbed again afterwards by something. I suspect this is because I have focus-follows-mouse turned on, and when the menu closes, the mouse ends up in a different window than the UIEditor.