Box/ValuePlugSerialiser/ExtensionAlgo : Improve export strategy
When exporting a Box for referencing or via ExtensionAlgo, we want to provide a way for the user to author default values for the reference/extension. In the past, this was done by transforming the current values into defaults at the point of export. This was error prone, as it was far too easy to change the values while testing your creation, and then accidentally export them as new defaults. Changing the defaults breaks compatibility with previous versions, so isn't to be taken lightly. We now export the current defaults unchanged and completely omit the current values. Instead, users can author defaults explicitly using the `ValuePlug::resetDefault()` method.