ImageViewUI : Small change which will be squashed in, presented separately due to relationship to crash
This change avoids the segfault which has been causing me problems - however I do not believe it is actually a solution.  It's slightly clearer code in this case anyway, but I don't think it should be illegal to call plug.parent() in a signal.  It causes a crash here because this signal is getting called during ImageView's destructor, and sometimes childPlug.parent() is referring to the ImageView itself, and constructor a new smart pointer to ImageView while it's destructing causes a double delete.  The next commit proposes a solution to avoid callbacks being called at all during ImageView's destruct, which makes this unnecessary ( though I'll still merge it anyway since it's cleaner ).