GraphGadget : Release GIL during construction
You would hope this would fix the hang in the test case, but it doesn't. It reveals that we _also_ need to release the GIL in the constructor for StandardNodeGadget. The tricky thing here is getting the `self` argument into the wrapper constructor, which isn't possible directly via `make_constructor`. This page suggests it can be done via `make_function` instead :

https://stackoverflow.com/questions/36200333/boost-python-return-same-instance-with-make-constructor

But then we have the question of when do we release the GIL? If we release it before calling the Wrapper constructor, then it is manipulating `self` without the GIL. No good. It seems we want to actually release in RefCountedWrapper where we call the constructor for `T`, but I don't know how to release the GIL around a call to a base class constructor. Are we stuck?

The real answer here may be that no UI should launch computes during construction, and that all computes should be done in the background.