Cycles : Move intensity/exposure/color handling to Renderer backend
Previously these were immune to ShaderTweaks because they were baked
into the `strength` parameter and `emission` shader by the CyclesLight
node. With all the post-processing to convert to `strength/emission`
moved to the CyclesRenderer, users now have a consistent view of their
lights which is editable at any point in the node graph.

Since CyclesLight no longer requires special logic, we can simplify the
implementation considerably, so that an internal `Shader` node does most
of the work, as it does in our other Light subclasses.

At this point I have removed the `image` parameter, on the assumption
that users can connect an `image_texture` node in manually if they want,
as you would do in Arnold. We can revisit this in future if necessary,
now we have more solid foundations, and even the odd unit test.