CatalogueUI : Use PathColumn event handling for output index column
This allows us encapsulate all the handling for the output index into the __OutputIndexColumn class, rather than having display code in one place and editing scattered through another class. This also adds a hover state to the column to indicate to users that there is something to be clicked on.