ArnoldRenderTest : Fix `testWedge()` by rendering to EXR
This was failing on L304, where the threshold for our "empty" render was being crossed, as follows :

- Arnold renders a watermark over the empty image, because there is no license server on GitHub CI. This is the reason for having a threshold in the first place, rather than testing against black.
- The watermark has pixels which have an alpha value lower than their RGB values. When unpremultiplied to perform a color transform in the ImageReader, this gives pixel values as high as 67.
- The ACES transform from `sRGB - Texture` to `ACEScg` applied by the ImageReader boosts these high values even further, to a max of 13924. When premultiplied again these are still as high as 54, blowing us well past our threshold.

What has changed to cause this to fail is the OCIO config. The old config didn't blow out highlights in its `sRGB` to `linear` transform at all - the max value was 1.3. The workaround is simple : render to an EXR file, so no colour transform is needed.