VisibleSet : Return `VisibleSet::Result` from `match()`
It was proving cumbersome to fit all of the specifics of a VisibleSet match into an `IECore::PathMatcher::Result` so we now replace it with a `VisibleSet::Result` struct.

This allows us to separate the path's `drawMode`:

- `None` : Should not be drawn.
- `Visible` : Should be drawn.
- `ExcludedBound` : Should be drawn but only as a bounding box signifying that the path has been excluded.

From whether the path could have `descendantsVisible`, which is now used to instruct the RenderController as to whether the path should be expanded.