RenderController : Use `VisibleSet::Result` to control visibility
`VisibleSet::Result::drawMode` informs us of the draw mode of a particular path.

Paths with a `drawMode` of `None` are not drawn, but may still be expanded via `VisibleSet::Result::descendantsVisible` in order to allow one or more of their descendants to be drawn.
Paths with a `drawMode` of `Visible` are drawn.
Paths with a `drawMode` of `ExcludedBounds` are only drawn as a Placeholder bounding box.