Dispatcher : Remove `nodes` arguments from signals
These are redundant, as the `tasks` plug on the dispatcher now provides access to the tasks being dispatched. The `nodes` argument was always a bit weird, because it only contained the first level of dispatched nodes, so you had to do your own traversal if you wanted to find all the upstream nodes anyway. And it's easier to do that starting from the `tasks` plug than it is from a list of nodes.