Dispatcher : Remove `script` and `context` arguments from `frameRange()`
The `context` argument was completely misplaced in the public API, where `Context::current()` is always passed implicitly.

The `script` argument was used for the `FullRange` mode, but doesn't make
much sense now dispatchers are typically parented to a script - it just makes it possible to pass the _wrong_ script. It was also dubious since the full frame range is also available from the context anyway, so passing a context _and_ a script makes it ambiguous as to where the full range should come from.

So we now just use the `frameRange:start` and `frameRange:end` variables from the current context. Those are the same as the script's frame range if the script's context is used, but can also be overridden dynamically. So for instance, you could now Wedge a bunch of shots with different full frame ranges by placing a ContextVariables node downstream of a dispatcher.