WIP
The problem we're trying to deal with here is that selection is retained even when the paths don't exist any more. So a newly created job can be selected just by virtue of a job being selected before. So we need to pretend selection has changed when a job is added, so that we connect to it if it _is_ selected.

When deleting a job, selection will jump randomly to the previous job in the pool, because job paths are based on index (and therefore aren't even stable). This almost looks intentional when the listing isn't sorted, but is completely random when it is. The only solution to that would be to add `PathListingWidget.pathAbove()` to find the best path to reselect, but even then I don't know how that would work with multiple selection (pass a list of paths?). And `pathAbove()` would require that we can get a job from a path string stably.