MultiLineTextWidget : Stop spurious `editingFinishedSignal()` emissions
Qt gives us the FocusOut event when a popup menu is shown for the widget, even though focus returns to the widget as soon as the menu is closed, and in the meantime the widget still `hasFocus()`. Since the user is often using a popup menu to modify the text, this leads to `editingFinishedSignal()` being emitted when editing definitely has not finished. Which means things like the Python Expression editor try to apply changes that are potentially not complete yet.

I've tested this fairly thoroughly on Linux, but we should probably test on other platforms too. It's not clear to me how platform-specific Qt's FocusOut behaviour is.

This change has consequences for any popup menu items for MultiLineStringPlugValueWidget. Previously, because `editingFinishedSignal` was emitted when the menu was shown, the text was copied to the plug before the menu was shown. Not it isn't. So menu items that want to modify the value should now source the start value from the widget's text and not from the plug value.

THIS BREAKS THE "COPY VALUE" MENU ITEM, BECAUSE IT IS A GENERIC THING THAT SOURCES FROM THE PLUG VALUE.