InteractiveRenderTest : Make test less sensitive
It seems that a PxrSphereLight gives out a _tiiiiny_ amount of light in channels where it has been told to be black. I don't know why. But that's not our concern here - we just want to check that the light gets added to the renderer, and _any_ illumination is sufficient to show that.