Light : Orient PxrDomeLight correctly
This matches the renderer to the default visualisation in Gaffer, which in turn matches the UsdLuxDomeLight default. This is the opposite of what we usually do, which is to adjust the visualisation to match. But this way a default-created PxrDomeLight is the right way up in Gaffer. And it seems that hdPrman makes a similar correction, suggesting that we need to do things this way to match eventual USD exports.