Material : Fix connections to OSL shader parameters
To make a connection in the Riley API, we need to know the type of the parameter. We were using the `.args` files to discover those, but since writing the code originally and now, all the pattern shaders have been converted to OSL, meaning they no longer have args files. We now use the type of the parameter based on the value found in the `Shader::parametersData()`. This does mean you need to supply the value as well as the connection, but Gaffer always does that anyway. We just need to make sure the unit tests do.

Also fixed a crash caused by throwing an exception during shader conversion. This would cause us to have a null material in the cache, which caused a null dereference in `MaterialCache::clearUnused()`.