IECoreRenderMan StaticTransform : Remove default constructor argument
This was dangerous, because the StaticTransform could continue to reference it after destruction. In practice, the existing usages of the default were safe because C++ lifetime rules dictated that the temporary for the default argument lived as long as the outer function call (the "full expression"). But it is too much of a pitfall to keep, and `IdentityTransform` is more descriptive for this use case anyway.