RendererMan : Define default custom LPE lobes
As mentioned in the comments, the RenderMan documentation lists defaults for these that don't actually exist. And the options aren't defined in `PRManOptions.args` either.

Our usual mantra is "expose the renderer directly", in which case we shouldn't be trying to provide defaults at all. But in this case I think it's the lesser of the evils :

- If we don't define a default, it's a real faff to set up denoising, because all the documentation for it says to use `U6`, which isn't defined by default.
- Likewise the documented LPEs for separate diffuse and subsurface etc rely on these default definitions.
- If we don't define a default, then the defaults come from `rendermn.ini`, which as installed defines U2 but nothing else. If we allow `rendermn.ini` to be in charge, then it would be really hard to show the defaults (we'd have to parse all the `.ini` files manually to figure it out).