Instancer : Use a node id instead of memory address
This is a more reliable way to uniquely identify nodes. The previous approach led to the risk of inconsistent results if a new node got the same address as a recently freed node ( this was probably quite unlikely in practical use, but appears to have become an issue in some tests ).