RenderMan Renderer : Support outputs required by OutputBuffer
This makes it possible to use RenderMan in Gaffer's 3D Viewer.

There are a couple of warnings from RenderMan when doing this :

```
R09073 {WARNING} Integer AOV 'id' should use only zmin or zmax filtering.
R09067 {WARNING} Inconsistent filter on 'z' while filter importance sampling.
```

The first suggests that we might want to allow filters to be specified per-output after all, although we could just hardcode to `zmin` for integers.

The second is baffling because we've definitely declared the `z` outputs to use the same global filter as everything else.