Wrapper : Move RenderMan configuration to Python
Maintaining separate wrappers for Windows and Linux is becoming a real pain, so I intend to refactor so that the platform-specific wrappers contain only the  bare minimum needed to launch Python, and then everything else is done in a Python-based wrapper. One lot of code in a language we are all proficient in - much better.

For now, just moving the RenderMan configuration to the Python wrapper, to minimise potential disruption in the 1.5.x series. But for 1.6 I intend to move everything into the Python wrapper.