Wrapper : Require `GAFFERENDERMAN_FEATURE_PREVIEW` to enable RenderMan
During the testing phase, we want folks to explicitly opt-in to testing, and we don't want any possibility of folks who haven't opted in being affected in any way. If the environment variable isn't specified, we don't even add GafferRenderMan to the module or library paths.

This also fixes a problem with the GafferScene startup files - if `RMANTREE` was set but GafferRenderMan hadn't been built, we would get an uncaught `ImportError`.