Light : Add `attributes` plug
This makes it easy for any derived class to specify arbitrary attributes for the light, without needing to override `hash/computeAttributes()`. The functionality is lifted up from OSLLight, so that it can be used everywhere.

This does leave us in a slighly untidy situation for now, because the `visualiserAttributes` plug is functionally identical but with a name that makes it inappropriate for wider use. And `mute` could just be a child of `attributes`. The slightly tricky part there is that we do want to separate the visualiser attributes in the UI, but that should be doable with a bit of metadata and a custom widget. We leave that for a slightly bigger refactoring of Light currently scheduled for Gaffer 1.6.