LocalDispatcher, SystemCommand, `gaffer env` : Preserve env-var case on Windows
On Windows, environment variable names are case-preserving but case-insensitive.
Python's `os.environ` mapping has a botched emulation for this which implements
the case-insensitive part by trashing the case and making everything upper-case.
See https://bugs.python.org/issue28824.

This caused us problems where we were using `os.environ.copy()` as the source for
a child environment we constructed for use with `subprocess`. The child environment
now had all-upper-case environment variables instead of the originals. And that caused
problems with `Context.substitute( "${var}" )` in the child process, because context
variable lookups are case-sensitive.

In the absence of a fix to Python itself, for now we add a `Gaffer.environment()`
function that builds a dictionary from the ground truth of the process' environment
table.

Fixes #6371.