Shader : Simplify connection handling
Our `effectiveParameter()` utility was used to find _any_ source parameter, be it an input or an output. But we no longer need to find _input_ parameters, because the previous commit switched to using Gaffer's standard compute mechanism for doing pass-throughs of parameter values from inputs to outputs.

All we need now is a way of finding source _output_ parameters to make connections from, which is what the refactored `connectionSource()` function provides. This removes a bunch of `isOutputParameter()` checks and generally simplifies everything.