Path : Remove `inspectionContext()`
It was always a bit awkward having this on the base class when it isn't relevant to most subclasses, and the only client is currently private. Now Context is RunTimeTyped we can use a property instead.

This opens up a possibility which will soon be useful : choosing between multiple properties to provide the context.