Path : Remove `inspectionContext()`
It was always a bit awkward having this on the base class when it isn't relevant to most subclasses, and the only client is currently private. Now Context is RunTimeTyped we can use a property instead.

It's worth noting that I've also moved responsibility for managing a canceller to the client : the context is no longer added by the paths. I think this makes sense since we're returning a potentially long-lived context, and there is no lifetime guarantee at all for the canceller passed to `property()`. It's also less of a worry than it was when we first added `inspectionContext()` - there are far fewer direct clients now that most clients just call `InspectorColumn::inspect()` and let it do everything.

This opens up a possibility which will soon be useful : choosing between multiple properties to provide the context.