RenderMan : Work around differences in XPU light linking behaviour
In XPU, a light with empty "grouping:membership" will contribute illumination to all objects, even those with a specific "lighting:subset". This is in contrast to RIS, which does the sensible thing whereby only lights explicitly mentioned by "lighting:subset" are used. It's not clear if this will be fixed on the RenderMan side, so for now at least we work around it by defaulting all lights to be in "defaultLightGroup".

Also replaced `m_extraAttributes` with `m_allAttributes`, storing the fully resolved attributes rather than reconstructing them each time they are need. This is less error-prone, and also minimises RtParamList shuffling in `updateLinking()`.