RenderMan : Fix XPU crash editing portal lights
The problem was that we were assigning `m_lightShader` before calling `ModifyLightInstance()` to modify the light to use the new shader. This meant we deleted the old light shader before the call to `ModifyLightInstance()`, putting the light in an invalid state - pointing to a deleted shader. XPU presumably accesses the current shader from `ModifyLightInstance()` since it crashes when called in this state.

This issue only affected portal lights, because all other light types use cached shaders, so the deletion of the old shader was deferred until `MaterialCache::clearUnused()`. We want to refactor so that portals use cached shaders too (see comment in `MaterialCache::getLightShader()`), but that is a bigger endeavour. In any case, it makes sense to do things by the book rather than rely on implementation details of the caching.