CustomOptionsUI, AttributeTweaksUI : Use customised PlugCreationWidget
This entails adding a signal to PlugCreationWidget to allow the menu to
be customised, so we can add items for each attribute found in the scene.
I added the menu items in CustomAttributesUI, using `ui:scene:acceptsAttributes`
metadata to trigger their creation, so that they are now available on CustomAttributes
as well as AttributeTweaks. The menu code is derived from the now-deleted AttributeTweaksUI
but the changes were big enough that the diffs weren't particularly readable. Hence I moved
it to CustomOptionsUI in one go, since that's where our other attribute-sensitive menus are.

This does remove the special handling of the `AttributeTweaks.localise` plug - we now always
the full list of inherited attributes unconditionally. But I think this makes sense even for
AttributeTweaks when localise is off, because Create mode tweaks don't care if the attribute
exists locally or not.